requirements.txt
flask==3.1.0
pandas==2.2.3
openpyxl==3.1.0
python-docx==1.1.2
pdfplumber==0.11.4
PyYAML==6.0
